shared_vk: report what the device can do, not what was asked for
Two claims in the shared Vulkan half were assertions rather than facts, and both would have been discovered on somebody else's machine.

The API version published was the INSTANCE's, which wgpu fills from the loader (vkEnumerateInstanceVersion) and then pins to 1.3 whenever the loader is 1.1 or better. What decides whether vkCmdPipelineBarrier2 exists is the PHYSICAL DEVICE's apiVersion, and ash fills an unloadable entry point with a stub that panics at the first call - inside whichever library called it. Both numbers are published now, and usable_api_version() is the lower of the two: the one a consumer must gate on.

synchronization2 was reported as true unconditionally, because this module chains the feature struct on. A driver may ignore a chain entry it does not recognise and still create the device, so on anything below 1.3 without VK_KHR_synchronization2 the flag claimed a feature nothing enabled - and every 2-form barrier a decoder records against such a device is undefined behaviour. The struct is now chained only where the feature can really exist, the extension is requested when it is the extension that provides it, and the flag is read back from the device.

The live test pins both as implications, so it fails on the machine where the lie would have mattered rather than passing everywhere.

+++ src/limits.rs
@@ -115,7 +115,8 @@ impl GpuLimits {
     ) -> Self {
         Self::new(
             limits.max_texture_dimension_2d,
-            u64::from(limits.max_storage_buffer_binding_size),
+            // Already `u64` in wgpu 30; it was `u32` when this was written.
+            limits.max_storage_buffer_binding_size,
             total_memory,
             available_memory,
             true,
